WEB/#57: Replaced Custom Design offer by Sense and Respond.
The idea here is to simplify the offer and make it more clear, easy to
understand and explain.

Custom design was some how similar to build and connect, and can cover
all the technical stack.

Sense and respond is easy to understand, if we combine it with
connectivity and cloud it can offer us an interesting niche.