WEB/#69: Added support for new field `message` to store user message into DB.
The new field is filled by user on text area and is sent with other POST data
to the php script `db_store.php`.

The `Send Message` button was changed to send data to the `db_store.php`
script.

Missing data like phone number was added to the form.

User message lengh is now displayed and bounded to 255 characters max.

Closes #69.