WEB/#102: Harmonized font size on Hero section badges.
Created a new HTML tag `HeroBadge` and used it into all hero sections.

All badges in all hero sections look the same except for the first page badge
which uses a darker color due to the dark background of the hero section on
that page.

Fixed small issue in make file.

Closes #102.